Moved functions specific to test into test package
