+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/impl/AbstractBotAuthenticatorTest.java
@@ -1,21 +1,25 @@
 package com.symphony.bdk.core.auth.impl;
 
+import static com.symphony.bdk.core.test.BdkRetryConfigTestHelper.ofMinimalInterval;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
 import com.symphony.bdk.core.auth.AuthSession;
 import com.symphony.bdk.core.auth.exception.AuthUnauthorizedException;
 import com.symphony.bdk.core.config.model.BdkRetryConfig;
 import com.symphony.bdk.http.api.ApiClient;
 import com.symphony.bdk.http.api.ApiException;
-
 import com.symphony.bdk.http.api.ApiRuntimeException;
 
 import org.junit.jupiter.api.Test;
 
 import javax.ws.rs.ProcessingException;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
 
 public class AbstractBotAuthenticatorTest {
 
@@ -39,7 +43,7 @@ public class AbstractBotAuthenticatorTest {
   void testSuccess() throws ApiException, AuthUnauthorizedException {
     final String token = "12324";
 
-    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(ofMinimalInterval()));
     doReturn(token).when(botAuthenticator).authenticateAndGetToken(any());
 
     assertEquals(token, botAuthenticator.retrieveToken(null));
@@ -48,7 +52,7 @@ public class AbstractBotAuthenticatorTest {
 
   @Test
   void testUnauthorized() throws ApiException {
-    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(401, "")).when(botAuthenticator).authenticateAndGetToken(any());
 
     assertThrows(AuthUnauthorizedException.class, () -> botAuthenticator.retrieveToken(null));
@@ -57,7 +61,7 @@ public class AbstractBotAuthenticatorTest {
 
   @Test
   void testUnexpectedApiException() throws ApiException {
-    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(404, "")).when(botAuthenticator).authenticateAndGetToken(any());
 
     assertThrows(ApiRuntimeException.class, () -> botAuthenticator.retrieveToken(null));
@@ -68,7 +72,7 @@ public class AbstractBotAuthenticatorTest {
   void testShouldRetry() throws ApiException, AuthUnauthorizedException {
     final String token = "12324";
 
-    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(BdkRetryConfig.ofMinimalInterval(4)));
+    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(ofMinimalInterval(4)));
     doThrow(new ApiException(429, ""))
         .doThrow(new ApiException(503, ""))
         .doThrow(new ProcessingException(""))
@@ -81,7 +85,7 @@ public class AbstractBotAuthenticatorTest {
 
   @Test
   void testRetriesExhausted() throws ApiException {
-    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(BdkRetryConfig.ofMinimalInterval(2)));
+    AbstractBotAuthenticator botAuthenticator = spy(new TestBotAuthenticator(ofMinimalInterval(2)));
     doThrow(new ApiException(429, "")).when(botAuthenticator).authenticateAndGetToken(any());
 
     assertThrows(ApiRuntimeException.class, () -> botAuthenticator.retrieveToken(null));
