+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/impl/AbstractExtensionAppAuthenticatorTest.java
@@ -1,8 +1,14 @@
 package com.symphony.bdk.core.auth.impl;
 
+import static com.symphony.bdk.core.test.BdkRetryConfigTestHelper.ofMinimalInterval;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.anyString;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 import com.symphony.bdk.core.auth.AppAuthSession;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
@@ -12,7 +18,6 @@ import com.symphony.bdk.core.config.model.BdkRetryConfig;
 import com.symphony.bdk.gen.api.model.ExtensionAppTokens;
 import com.symphony.bdk.gen.api.model.PodCertificate;
 import com.symphony.bdk.http.api.ApiException;
-
 import com.symphony.bdk.http.api.ApiRuntimeException;
 
 import org.junit.jupiter.api.Test;
@@ -53,7 +58,7 @@ public class AbstractExtensionAppAuthenticatorTest {
     final String certificate = "cert";
 
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doReturn(new PodCertificate().certificate(certificate)).when(authenticator).callGetPodCertificate();
 
     assertEquals(certificate, authenticator.getPodCertificate().getCertificate());
@@ -63,7 +68,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetPodCertificateUnauthorized() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(401, "")).when(authenticator).callGetPodCertificate();
 
     assertThrows(ApiRuntimeException.class, () -> authenticator.getPodCertificate().getCertificate());
@@ -73,7 +78,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetPodCertificateUnexpectedApiException() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(404, "")).when(authenticator).callGetPodCertificate();
   }
 
@@ -82,7 +87,7 @@ public class AbstractExtensionAppAuthenticatorTest {
     final String certificate = "cert";
 
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(429, ""))
         .doThrow(new ApiException(503, ""))
         .doThrow(new ProcessingException(""))
@@ -96,7 +101,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetPodCertificateRetriesExhausted() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval(2)));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval(2)));
     doThrow(new ApiException(503, "")).when(authenticator).callGetPodCertificate();
 
     assertThrows(ApiRuntimeException.class, () -> authenticator.getPodCertificate());
@@ -110,7 +115,7 @@ public class AbstractExtensionAppAuthenticatorTest {
         new ExtensionAppTokens().appId("appId").appToken("Ta").symphonyToken("Ts");
 
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doReturn(extensionAppTokens).when(authenticator).authenticateAndRetrieveTokens(anyString());
 
     assertEquals(extensionAppTokens, authenticator.retrieveExtAppTokens(""));
@@ -120,7 +125,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetTokensUnauthorized() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(401, "")).when(authenticator).authenticateAndRetrieveTokens(anyString());
 
     assertThrows(AuthUnauthorizedException.class, () -> authenticator.retrieveExtAppTokens(""));
@@ -130,7 +135,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetTokensUnexpectedApiException() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(404, "")).when(authenticator).authenticateAndRetrieveTokens(anyString());
 
     assertThrows(ApiRuntimeException.class, () -> authenticator.retrieveExtAppTokens(""));
@@ -143,7 +148,7 @@ public class AbstractExtensionAppAuthenticatorTest {
         new ExtensionAppTokens().appId("appId").appToken("Ta").symphonyToken("Ts");
 
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval()));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval()));
     doThrow(new ApiException(429, ""))
         .doThrow(new ApiException(503, ""))
         .doThrow(new ProcessingException(""))
@@ -157,7 +162,7 @@ public class AbstractExtensionAppAuthenticatorTest {
   @Test
   void testGetTokensRetriesExhausted() throws ApiException {
     AbstractExtensionAppAuthenticator authenticator =
-        spy(new TestExtAppAuthenticator(BdkRetryConfig.ofMinimalInterval(2)));
+        spy(new TestExtAppAuthenticator(ofMinimalInterval(2)));
     doThrow(new ApiException(429, "")).when(authenticator).authenticateAndRetrieveTokens(anyString());
 
     assertThrows(ApiRuntimeException.class, () -> authenticator.retrieveExtAppTokens(""));
