+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/impl/AuthenticationRetryTest.java
@@ -1,26 +1,28 @@
 package com.symphony.bdk.core.auth.impl;
 
+import static com.symphony.bdk.core.test.BdkRetryConfigTestHelper.ofMinimalInterval;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.when;
+
 import com.symphony.bdk.core.auth.exception.AuthUnauthorizedException;
-import com.symphony.bdk.core.config.model.BdkRetryConfig;
-
 import com.symphony.bdk.core.util.function.SupplierWithApiException;
-
 import com.symphony.bdk.http.api.ApiException;
-
 import com.symphony.bdk.http.api.ApiRuntimeException;
 
 import org.junit.jupiter.api.Test;
 
 import javax.ws.rs.ProcessingException;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
 class AuthenticationRetryTest {
 
   @Test
   void testCallSucceedsFirstTime() throws AuthUnauthorizedException {
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(1));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(1));
 
     final String output = "output";
     assertEquals(output, authenticationRetry.executeAndRetry("test", () -> output, ""));
@@ -33,7 +35,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ApiException(502, "")).thenReturn(output);
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
     assertEquals(output, authenticationRetry.executeAndRetry("test", supplier, ""));
 
     verify(supplier, times(2)).get();
@@ -47,7 +49,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ApiException(429, "")).thenReturn(output);
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertEquals(output, authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(2)).get();
@@ -61,7 +63,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ProcessingException("")).thenReturn(output);
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertEquals(output, authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(2)).get();
@@ -73,7 +75,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ApiException(429, ""));
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertThrows(ApiRuntimeException.class, () -> authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(2)).get();
@@ -85,7 +87,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ApiException(401, ""));
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertThrows(AuthUnauthorizedException.class, () -> authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(1)).get();
@@ -97,7 +99,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new ApiException(404, ""));
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertThrows(ApiRuntimeException.class, () -> authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(1)).get();
@@ -109,7 +111,7 @@ class AuthenticationRetryTest {
     SupplierWithApiException<String> supplier = mock(SupplierWithApiException.class);
     when(supplier.get()).thenThrow(new IllegalStateException(""));
 
-    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(BdkRetryConfig.ofMinimalInterval(2));
+    AuthenticationRetry<String> authenticationRetry = new AuthenticationRetry<>(ofMinimalInterval(2));
 
     assertThrows(RuntimeException.class, () -> authenticationRetry.executeAndRetry("test", supplier, ""));
     verify(supplier, times(1)).get();
